third_party/blink/web_tests/platform/generic/fast/js/mozilla/strict/11.4.4-expected.txt

PASS Function("'use strict'; ++arguments") threw exception of type SyntaxError.
PASS !!Function("++arguments") is true
PASS true === true
PASS Function("'use strict'; ++eval") threw exception of type SyntaxError.
PASS !!Function("++eval") is true
PASS true === true
PASS Function("'use strict'; ++(arguments)") threw exception of type SyntaxError.
PASS !!Function("++(arguments)") is true
PASS true === true
PASS Function("'use strict'; ++(eval)") threw exception of type SyntaxError.
PASS !!Function("++(eval)") is true
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
